Improve projection structure to make it more "autowireable"
