+++ src/Infrastructure/MongoDB/DroppableRepositoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace WalletAccountant\Infrastructure\MongoDB;
+
+/**
+ * DroppableRepositoryInterface
+ */
+interface DroppableRepositoryInterface
+{
+    public function dropCollection(): void;
+}
